Rename partcad-client-utils to partcad-client
The package is the PartCAD client -- finding and talking to the daemon
serving a workspace, and upgrading this installation -- not a bag of
utilities beside one. `partcad-utils` keeps the "-utils" name because
that is what it is.

A pure rename: the distribution, the directory, and the Python package
(`partcad_client_utils` -> `partcad_client`). No shims, since nothing has
been released under the old name. Every reference moves with it --
imports, the Poetry package list and pytest path, the build/deploy/test
workflows, the PyInstaller spec and build script, bumpversion, gitignore,
and the prose in every AGENTS.md and README.

Also drops the last "utilities" wording from the package's own
description and README opening, which the rename made inaccurate.
