+++ .github/workflows/deploy.yml
@@ -64,12 +64,12 @@ jobs:
       - name: Test building and packaging
         run: |
           # Build the shared partcad-utils wheel first; partcad and the CLI/service
-          # pin it, and it is not published on PyPI. partcad-client-utils pins it
+          # pin it, and it is not published on PyPI. partcad-client pins it
           # in turn, and partcad-cli pins that.
           cp LICENSE.txt partcad-utils
           (. .venv/build/${{ env.BIN_DIR }}/activate && cd partcad-utils && python -m build && cd .. && deactivate)
-          cp LICENSE.txt partcad-client-utils
-          (. .venv/build/${{ env.BIN_DIR }}/activate && cd partcad-client-utils && python -m build && cd .. && deactivate)
+          cp LICENSE.txt partcad-client
+          (. .venv/build/${{ env.BIN_DIR }}/activate && cd partcad-client && python -m build && cd .. && deactivate)
           (. .venv/build/${{ env.BIN_DIR }}/activate && cd partcad && python -m build && cd .. && deactivate)
           (. .venv/build-cli/${{ env.BIN_DIR }}/activate && python -m pip install --find-links=partcad-utils/dist partcad-utils && deactivate)
           (. .venv/build-cli/${{ env.BIN_DIR }}/activate && python -m pip install -r partcad/requirements.txt && deactivate)
@@ -82,7 +82,7 @@ jobs:
       - name: Test installation
         run: |
           (. .venv/install/${{ env.BIN_DIR }}/activate && python -m pip install --find-links=partcad-utils/dist partcad-utils && deactivate)
-          (. .venv/install/${{ env.BIN_DIR }}/activate && python -m pip install --find-links=partcad-client-utils/dist --find-links=partcad-utils/dist partcad-client-utils && deactivate)
+          (. .venv/install/${{ env.BIN_DIR }}/activate && python -m pip install --find-links=partcad-client/dist --find-links=partcad-utils/dist partcad-client && deactivate)
           (. .venv/install/${{ env.BIN_DIR }}/activate && python -m pip install -r partcad/requirements.txt     && deactivate)
           (. .venv/install/${{ env.BIN_DIR }}/activate && python -m pip install --no-index --find-links=partcad/dist     partcad && deactivate)
           (. .venv/install/${{ env.BIN_DIR }}/activate && python -m pip install --no-index --find-links=partcad-service-json-rpc/dist partcad-service-json-rpc && deactivate)
@@ -96,7 +96,7 @@ jobs:
           name: python-package-distributions
           path: partcad/dist/
       # "partcad" pins "partcad-utils"; "partcad-cli" pins that,
-      # "partcad-client-utils" and "partcad-service-json-rpc". None of them
+      # "partcad-client" and "partcad-service-json-rpc". None of them
       # exists on PyPI yet, so all of them have to be published with this release
       # or the released wheels cannot be installed.
       - name: Upload "partcad-utils"
@@ -105,12 +105,12 @@ jobs:
         with:
           name: python-package-distributions-utils
           path: partcad-utils/dist/
-      - name: Upload "partcad-client-utils"
+      - name: Upload "partcad-client"
         if: ${{ matrix.os == 'ubuntu-24.04' && matrix.python-version == '3.13' }}
         uses: actions/upload-artifact@v7
         with:
           name: python-package-distributions-client-utils
-          path: partcad-client-utils/dist/
+          path: partcad-client/dist/
       - name: Upload "partcad-service-json-rpc"
         if: ${{ matrix.os == 'ubuntu-24.04' && matrix.python-version == '3.13' }}
         uses: actions/upload-artifact@v7
