+++ dev-tools/bumpversion.toml
@@ -24,7 +24,7 @@ replace = "partcad=={new_version}"
 # Every package in the monorepo, and every pin between them. `partcad-utils` and
 # `partcad-service-json-rpc` were once added without being added here, so their
 # versions stayed at whatever they were created with while everything else moved.
-# That is not cosmetic: `pc upgrade` compares `partcad_client_utils.__version__`
+# That is not cosmetic: `pc upgrade` compares `partcad_client.__version__`
 # against the newest release to decide whether to update, so a stale constant
 # means the CLI either never upgrades or upgrades forever.
 # `partcad-cli/tests/unit/test_versions.py` fails if any of these fall out of
@@ -40,24 +40,24 @@ search = "__version__ = \"{current_version}\""
 replace = "__version__ = \"{new_version}\""
 
 [[tool.bumpversion.files]]
-filename = "partcad-client-utils/pyproject.toml"
+filename = "partcad-client/pyproject.toml"
 search = "version = \"{current_version}\""
 replace = "version = \"{new_version}\""
 
 [[tool.bumpversion.files]]
-filename = "partcad-client-utils/src/partcad_client_utils/__init__.py"
+filename = "partcad-client/src/partcad_client/__init__.py"
 search = "__version__ = \"{current_version}\""
 replace = "__version__ = \"{new_version}\""
 
 [[tool.bumpversion.files]]
-filename = "partcad-client-utils/requirements.txt"
+filename = "partcad-client/requirements.txt"
 search = "partcad-utils=={current_version}"
 replace = "partcad-utils=={new_version}"
 
 [[tool.bumpversion.files]]
 filename = "partcad-cli/requirements.txt"
-search = "partcad-client-utils=={current_version}"
-replace = "partcad-client-utils=={new_version}"
+search = "partcad-client=={current_version}"
+replace = "partcad-client=={new_version}"
 
 [[tool.bumpversion.files]]
 filename = "partcad-service-json-rpc/pyproject.toml"
