+++ dev-tools/pyinstaller/partcad.spec
@@ -133,7 +133,7 @@ hiddenimports += command_modules()
 # import them directly (some, like the telemetry backends, only by name), so
 # collect the whole package.
 hiddenimports += collect_submodules("partcad_utils")
-hiddenimports += collect_submodules("partcad_client_utils")
+hiddenimports += collect_submodules("partcad_client")
 
 # The JSON-RPC service (`partcad-json-rpc`). Its HTTP transport imports aiohttp
 # lazily, inside a function that is only reached in HTTP mode, so PyInstaller's
@@ -200,7 +200,7 @@ else:
 # The version PartCAD reports and sends with telemetry.
 add_metadata("partcad")
 add_metadata("partcad-utils")
-add_metadata("partcad-client-utils")
+add_metadata("partcad-client")
 add_metadata("partcad-cli")
 add_metadata("partcad-service-json-rpc")
