+++ partcad-cli/tests/unit/test_versions.py
@@ -5,7 +5,7 @@
 #
 """The monorepo releases every component under one version. This checks it.
 
-`pc upgrade` asks `partcad_client_utils.__version__` what is installed and compares it
+`pc upgrade` asks `partcad_client.__version__` what is installed and compares it
 against the newest release. A version constant that stopped moving is therefore
 not cosmetic: it makes the CLI either never upgrade or upgrade forever. Two of
 them had stopped moving -- `partcad-utils` and `partcad-service-json-rpc` were
@@ -21,7 +21,7 @@ import tomllib
 from pathlib import Path
 
 import partcad_cli
-import partcad_client_utils
+import partcad_client
 import partcad_service_json_rpc
 import partcad_utils
 import pytest
@@ -31,12 +31,12 @@ import partcad
 REPO_ROOT = Path(__file__).resolve().parents[3]
 BUMPVERSION = REPO_ROOT / "dev-tools" / "bumpversion.toml"
 
-# Every package that reports a version at runtime. `partcad_client_utils` is the
+# Every package that reports a version at runtime. `partcad_client` is the
 # one `pc upgrade` actually reads.
 PACKAGES = {
     "partcad": partcad,
     "partcad-cli": partcad_cli,
-    "partcad-client-utils": partcad_client_utils,
+    "partcad-client": partcad_client,
     "partcad-utils": partcad_utils,
     "partcad-service-json-rpc": partcad_service_json_rpc,
 }
