+++ partcad-service-json-rpc/README.md
@@ -43,12 +43,12 @@ Call `rpc.discover` to list the available methods and their summaries.
 
 Not here, deliberately. Updating an installation is a client-side act — it is the client's machine, and its copy
 of PartCAD, that gets replaced — and a daemon can be remote, where "update PartCAD" would mean updating somebody
-else's installation. The implementation is `partcad_client_utils.selfupdate`; the way to run it is `pc upgrade`,
+else's installation. The implementation is `partcad_client.selfupdate`; the way to run it is `pc upgrade`,
 which stops every daemon running on the machine and waits for them before installing. This service neither
 updates itself nor goes looking for other daemons to stop: doing it from the client means one process acting on
 its own machine, rather than daemons racing each other.
 
-Finding and stopping daemons is `partcad_client_utils.daemon` for the same reason. What is here is the serving
+Finding and stopping daemons is `partcad_client.daemon` for the same reason. What is here is the serving
 half — `daemon.ensure_daemon` and the transports. The address they bind, and the framing they speak, are
 `partcad_utils.workspace` and `partcad_utils.framing`, shared with clients so the two ends cannot disagree
 about where to meet.
